+++ docs/changelog.md
@@ -1,11 +1,22 @@
 # TBD
+### Changes
+- Prepared the Kurtosis engine server to do search in logs
+- Adding `log line filters` parameter in the `GetServiceLogs` Kurtosis engine endpoint
+
+### Fixes
+- Fixed a bug which was happening on small terminal windows regarding the display of the progress bar and progress info
+
+# 0.57.6
 ### Features
 - The "Starlark code successfully executed" or "Error encountered running Starlark code" messages are now "Starlark 
 code successfully run in dry-run mode" and "Error encountered running Starlark code in dry-run mode" when Starlark is 
-run in dry-run mode (and without the "in dry-run mode" when the script is executed for real)
+run in dry-run mode (and without the "in dry-run mode" when the script is executed for real)\
+- Added `RunStarlarkScriptBlocking`, `RunStarlarkPackageBlocking` and `RunStarlarkRemotePackageBlocking` functions
+to the enclave context to facilitate automated testing in our current modules.
 
 ### Fixes
 - Don't duplicate instruction position information in `store_service_files`
+- Use constants instead of hardcoded string for validation errors
 
 ### Removals
 - Remove stack trace from validation errors as it isn't used currently
@@ -13,8 +24,6 @@ run in dry-run mode (and without the "in dry-run mode" when the script is execut
 ### Changes
 - Changed validation message from "Pre-validating" to "Validating"
 - Disabled progress info in non-interactive terminals when running a Starlark Package
-- Prepared the Kurtosis engine server to do search in logs
-- Adding `log line filters` parameter in the `GetServiceLogs` Kurtosis engine endpoint
 
 # 0.57.5
 ### Changes
@@ -60,6 +69,7 @@ run in dry-run mode (and without the "in dry-run mode" when the script is execut
 
 ### Features
 - Log file name and function like [filename.go:FunctionName()] while logging in `core` & `engine`
+- Add artifact ID validation to Starlark commands
 - Add IP address string replacement in `print` command
 - All Kurtosis instructions now returns a simple but explicit output
 - The object returned by Starlark's `run()` function is serialized as JSON and returned to the CLI output.